requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
cupy-cuda112==10.1.0
cycler==0.11.0
docopt==0.6.2
fastrlock==0.8
fonttools==4.29.1
idna==3.3
kiwisolver==1.3.2
llvmlite==0.38.0
matplotlib==3.5.1
numba==0.55.1
numpy==1.21.5
opencv-python==4.5.5.62
packaging==21.3
Pillow==9.0.0
pipreqs==0.4.11
pyparsing==3.0.7
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
urllib3==1.26.8
yarg==0.1.9
